django_images/requirements.txt
Django==4.0
gunicorn==20.1.0
channels==4.0
celery==5.2.3
pillow==9.1.1
redis==4.1.4
pika==1.2.0 
psycopg2-binary==2.9.3  
